changes simulator test( has error )

+++ src/instruction.h
@@ -3,6 +3,6 @@
 #include <stdint.h>
 
 int isOpcodePrefix(uint8_t *code);
-uint8_t instruction(uint8_t **codePtr);
+int instruction(uint8_t **codePtr);
 
 #endif // _INSTRUCTION_H
